Adding CRUD Functionality for Coin Collection Management (#1)
* Initial commit

* Added settings for the database and specified the server port
This commit includes adding configuration files for the database, as well as specifying the server port in the configuration. The application is now ready to connect to the database and the server will listen on the specified port.

* A coin model has been created and the main fields have been entered.

* Added link fields to other models.

* The main models characterizing the state of Coin have been created.

* Add Condition enum and Price model with composite key
This commit introduces a Condition enum that represents different states a coin can have, providing a clear and standardized way to manage coin statuses.
Additionally, a Price model is created, featuring a composite key composed of coin_id and condition. This allows for a precise association between coin and condition when handling pricing information.

* Documented

* Refactored documentation

* Initial commit for repositories

* Change key type

* Fixed a stack overflow bug by excluding looped fields from the toString() method.

* Initial commit

* Initial commit, base methods

* Added new methods